split request and SQLite class implementation in files.
